TreeData type into shared file
