+++ packages/studio-base/src/components/TopicList/types.ts
@@ -1,0 +1,15 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+export type TreeData = {
+  id: string;
+  name: string;
+  aliasedFromName?: string;
+  schemaName?: string;
+  isComplex?: boolean;
+  isArray?: boolean;
+  isConstant?: boolean;
+  children?: TreeData[];
+  messagePath: string;
+};
